api/requirements.txt
Flask==3.0.0
supabase==2.0.2
python-dotenv==1.0.0
requests==2.31.0
python-dateutil>=2.8.0
APScheduler==3.10.4
beautifulsoup4
openai==1.6.1
flask-cors==4.0.0
